+++ tests/testthat/test-hermes.R
@@ -21,12 +21,14 @@ test_that("hermes download works", {
 
   path_hermes <- tempdir()
 
-  # Test path
-  expect_no_error(get_path_hermes_obj("objectname", "projectname", path_hermes))
+ # Define required variables
+ region <- Region(name = "nebraska", type = "us state",
+                  div = c(country = "United States", state = "Nebraska"))
 
-  # Save and load object
-  x <- 1:3
-  expect_no_error(save_obj(x, "objectname", "projectname", path_hermes))
-  expect_no_error(read_obj("objectname", "projectname", path_hermes))
+ expect_true(dir.exists(create_dir_hermes(region,
+                                          "cronus",
+                                          "plot",
+                                          "Progress",
+                                          dir = path_hermes)))
 
 })